Stats: Update stats as of 2022-08-16 12:23 (#182)
